requirements.txt
openai
pillow
pyyaml